Created an abstract Block, BlockType, removed encoding
