+++ checkov/common/graph/db_connectors/networkx/networkx_db_connector.py
@@ -19,7 +19,7 @@ class NetworkxConnector(DBConnector):
 
     def networkx_from_local_graph(self, local_graph):
         self.graph = nx.DiGraph()
-        vertices_attributes = [v.get_decoded_attribute_dict() for v in local_graph.vertices]
+        vertices_attributes = [v.get_attribute_dict() for v in local_graph.vertices]
         vertices_to_add = [(attr[CustomAttributes.HASH], attr) for attr in vertices_attributes]
         edges_to_add = [(vertices_attributes[e.origin][CustomAttributes.HASH], vertices_attributes[e.dest][CustomAttributes.HASH], {'label': e.label}) for e in local_graph.edges]
