+++ checkov/common/graph/graph_builder/graph_components/block_types.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class BlockType:
+    RESOURCE = "resource"
+
+    def get(self, attr_name):
+        return getattr(self, attr_name.upper())
