+++ checkov/terraform/graph_builder/graph_components/module.py
@@ -4,7 +4,8 @@ from typing import List, Dict, Any, Set, Callable
 
 from checkov.terraform.checks.utils.dependency_path_handler import unify_dependency_path
 from checkov.terraform.graph_builder.graph_components.block_types import BlockType
-from checkov.terraform.graph_builder.graph_components.blocks import Block, get_inner_attributes
+from checkov.terraform.graph_builder.graph_components.blocks import TerraformBlock
+from checkov.common.graph.graph_builder.graph_components.blocks import get_inner_attributes
 
 
 class Module:
@@ -18,7 +19,7 @@ class Module:
         self.dep_index_mapping = dep_index_mapping
         self.module_dependency_map = module_dependency_map
         self.path = ""
-        self.blocks: List[Block] = []
+        self.blocks: List[TerraformBlock] = []
         self.customer_name = ""
         self.account_id = ""
         self.source = ""
@@ -33,7 +34,7 @@ class Module:
         if self._block_type_to_func.get(block_type):
             self._block_type_to_func[block_type](self, blocks, path)
 
-    def _add_to_blocks(self, block: Block) -> None:
+    def _add_to_blocks(self, block: TerraformBlock) -> None:
         dependencies = [dep_trail for dep_trail in self.module_dependency_map.get(os.path.dirname(block.path), [])]
         module_dependency_num = ""
         if not dependencies:
@@ -54,7 +55,7 @@ class Module:
                 alias = attributes.get("alias")
                 if alias:
                     provider_name = f"{provider_name}.{alias[0]}"
-                provider_block = Block(
+                provider_block = TerraformBlock(
                     block_type=BlockType.PROVIDER,
                     name=provider_name,
                     config=provider_dict,
@@ -69,7 +70,7 @@ class Module:
         for variable_dict in blocks:
             for name in variable_dict:
                 attributes = variable_dict[name]
-                variable_block = Block(
+                variable_block = TerraformBlock(
                     block_type=BlockType.VARIABLE,
                     name=name,
                     config=variable_dict,
@@ -83,7 +84,7 @@ class Module:
     def _add_locals(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for blocks_section in blocks:
             for name in blocks_section:
-                local_block = Block(
+                local_block = TerraformBlock(
                     block_type=BlockType.LOCALS,
                     name=name,
                     config={name: blocks_section[name]},
@@ -99,7 +100,7 @@ class Module:
             for name in output_dict:
                 if type(output_dict[name]) is not dict:
                     continue
-                output_block = Block(
+                output_block = TerraformBlock(
                     block_type=BlockType.OUTPUT,
                     name=name,
                     config=output_dict,
@@ -113,7 +114,7 @@ class Module:
     def _add_module(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for module_dict in blocks:
             for name in module_dict:
-                module_block = Block(
+                module_block = TerraformBlock(
                     block_type=BlockType.MODULE,
                     name=name,
                     config=module_dict,
@@ -134,7 +135,7 @@ class Module:
                     if provisioner:
                         self._handle_provisioner(provisioner, attributes)
                     attributes["resource_type"] = [resource_type]
-                    resource_block = Block(
+                    resource_block = TerraformBlock(
                         block_type=BlockType.RESOURCE,
                         name=f"{resource_type}.{name}",
                         config=resource_dict,
@@ -150,7 +151,7 @@ class Module:
         for data_dict in blocks:
             for data_type in data_dict:
                 for name in data_dict[data_type]:
-                    data_block = Block(
+                    data_block = TerraformBlock(
                         block_type=BlockType.DATA,
                         name=data_type + "." + name,
                         config=data_dict,
@@ -165,7 +166,7 @@ class Module:
     def _add_terraform_block(self, blocks: List[Dict[str, Dict[str, Any]]], path: str) -> None:
         for terraform_dict in blocks:
             for name in terraform_dict:
-                terraform_block = Block(
+                terraform_block = TerraformBlock(
                     block_type=BlockType.TERRAFORM,
                     name=name,
                     config=terraform_dict,
@@ -178,7 +179,7 @@ class Module:
 
     def _add_tf_var(self, blocks: Dict[str, Dict[str, Any]], path: str) -> None:
         for tf_var_name, attributes in blocks.items():
-            tfvar_block = Block(
+            tfvar_block = TerraformBlock(
                 block_type=BlockType.TF_VARIABLE,
                 name=tf_var_name,
                 config={tf_var_name: attributes},
